Use uint32_t instaead of size_t to be not platform dependent in inter-communication + fix some logic bugs

+++ src/network/interCom/interMsg/interMsg.cpp
@@ -16,7 +16,7 @@ InterMsg::InterMsg(Client client,
 void InterMsg::send()
 {
     char l_data[InterMsgData::MAX_PRIV_DATA_SIZE];
-    size_t l_size = m_data.getPrivData(l_data);
+    uint32_t l_size = m_data.getPrivData(l_data);
     const char *l_ip = m_client.getIP().getIpString().c_str();
     switch (m_sendType)
     {
